+++ hutool-core/src/main/java/cn/hutool/core/util/RandomUtil.java
@@ -11,12 +11,12 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Random;
 import java.util.Set;
-import java.util.UUID;
 import java.util.concurrent.ThreadLocalRandom;
 
 import cn.hutool.core.date.DateTime;
 import cn.hutool.core.date.DateUtil;
 import cn.hutool.core.exceptions.UtilException;
+import cn.hutool.core.lang.UUID;
 import cn.hutool.core.lang.WeightRandom;
 import cn.hutool.core.lang.WeightRandom.WeightObj;
 
@@ -382,9 +382,9 @@ public class RandomUtil {
 	}
 
 	/**
-	 * 随机数字
+	 * 随机数字，数字为0~9单个数字
 	 * 
-	 * @return 随机字符
+	 * @return 随机数字字符
 	 * @since 3.1.2
 	 */
 	public static int randomNumber() {
@@ -460,7 +460,7 @@ public class RandomUtil {
 	 * @since 3.2.2
 	 */
 	public static String simpleUUID() {
-		return randomUUID().replace("-", "");
+		return UUID.randomUUID().toString(true);
 	}
 
 	/**