+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/rythm/RythmEngine.java
@@ -7,7 +7,8 @@ import cn.hutool.extra.template.Template;
 import cn.hutool.extra.template.TemplateConfig;
 
 /**
- * Rythm模板引擎
+ * Rythm模板引擎<br>
+ * 文档：http://rythmengine.org/doc/index
  * 
  * @author looly
  *
@@ -43,11 +44,6 @@ public class RythmEngine implements Engine {
 	}
 	// --------------------------------------------------------------------------------- Constructor end
 	
-	@Override
-	public String getName() {
-		return "Rythm";
-	}
-
 	@Override
 	public Template getTemplate(String resource) {
 		return RythmTemplate.wrap(engine.getTemplate(resource));